requirements.txt
pandas
snowflake-connector-python[pandas]>=2.7.9
tfcausalimpact
ipykernel
seaborn
pyyaml
black
